mc: refactor identity related stuffs out of util
