+++ mc/util.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	p2p_crypto "github.com/libp2p/go-libp2p-crypto"
 	p2p_host "github.com/libp2p/go-libp2p-host"
 	p2p_metrics "github.com/libp2p/go-libp2p-metrics"
 	p2p_peer "github.com/libp2p/go-libp2p-peer"
@@ -12,10 +11,6 @@ import (
 	p2p_swarm "github.com/libp2p/go-libp2p-swarm"
 	p2p_bhost "github.com/libp2p/go-libp2p/p2p/host/basic"
 	multiaddr "github.com/multiformats/go-multiaddr"
-	"io/ioutil"
-	"log"
-	"os"
-	"path"
 	"strings"
 )
 
@@ -62,81 +57,6 @@ func ParseHandleId(str string) (empty p2p_pstore.PeerInfo, err error) {
 	}
 
 	return p2p_pstore.PeerInfo{ID: pid}, nil
-}
-
-// node identities
-type Identity struct {
-	ID      p2p_peer.ID
-	PrivKey p2p_crypto.PrivKey
-}
-
-func (id Identity) Pretty() string {
-	return id.ID.Pretty()
-}
-
-func NodeIdentity(home string) (empty Identity, err error) {
-	kpath := path.Join(home, "identity")
-	_, err = os.Stat(kpath)
-	if os.IsNotExist(err) {
-		return generateIdentity(kpath)
-	}
-	if err != nil {
-		return
-	}
-	return loadIdentity(kpath)
-}
-
-func generateIdentity(kpath string) (empty Identity, err error) {
-	log.Printf("Generating new identity")
-	privk, pubk, err := generateKeyPair()
-	if err != nil {
-		return
-	}
-
-	id, err := p2p_peer.IDFromPublicKey(pubk)
-	if err != nil {
-		return
-	}
-
-	log.Printf("Saving identity to %s", kpath)
-	bytes, err := privk.Bytes()
-	if err != nil {
-		return
-	}
-
-	err = ioutil.WriteFile(kpath, bytes, 0600)
-	if err != nil {
-		return
-	}
-
-	log.Printf("ID: %s", id.Pretty())
-	return Identity{ID: id, PrivKey: privk}, nil
-}
-
-func loadIdentity(kpath string) (empty Identity, err error) {
-	log.Printf("Loading identity from %s", kpath)
-	bytes, err := ioutil.ReadFile(kpath)
-	if err != nil {
-		return
-	}
-
-	privk, err := p2p_crypto.UnmarshalPrivateKey(bytes)
-	if err != nil {
-		return
-	}
-
-	id, err := p2p_peer.IDFromPrivateKey(privk)
-	if err != nil {
-		return
-	}
-
-	log.Printf("ID: %s", id.Pretty())
-	return Identity{ID: id, PrivKey: privk}, nil
-}
-
-func generateKeyPair() (p2p_crypto.PrivKey, p2p_crypto.PubKey, error) {
-	// 1kbit RSA just for now -- until ECC key support is merged in libp2p
-	return p2p_crypto.GenerateKeyPair(p2p_crypto.RSA, 1024)
 }
 
 func NewHost(id Identity, addrs ...multiaddr.Multiaddr) (p2p_host.Host, error) {
